Rename Mapper => Dao, MappingManager => Mapper
