+++ mapper-runtime/src/main/java/com/datastax/oss/driver/api/mapper/annotations/Dao.java
@@ -13,9 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.datastax.oss.driver.mapper;
+package com.datastax.oss.driver.api.mapper.annotations;
 
-import com.datastax.oss.driver.api.mapper.annotations.Mapper;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
-@Mapper
-public interface ProductMapper {}
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.CLASS)
+public @interface Dao {}
