--- mapper-runtime/src/main/java/com/datastax/oss/driver/api/mapper/annotations/MappingManager.java
@@ -1,35 +1,0 @@
-/*
- * Copyright DataStax, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.datastax.oss.driver.api.mapper.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.CLASS)
-public @interface MappingManager {
-
-  /**
-   * The fully-qualified name of the builder class that will get generated in order to create
-   * instances of the manager, for example "com.mycompany.MyCustomBuilder".
-   *
-   * <p>If this is left empty (the default), the builder will use the name of the original interface
-   * with the suffix "Builder", and reside in the same package.
-   */
-  String builderName() default "";
-}
